refactor(front): Change naming for image inspector components
