+++ ui/components/imageWorkspace/src/components/ImageInspector/ArrowSvg.svelte
