+++ ui/components/imageWorkspace/src/components/ImageInspector/ObjectItem.svelte
