+++ ui/components/imageWorkspace/src/components/ImageInspector/ObjectsModelSection.svelte
