+++ ui/components/imageWorkspace/src/components/ImageInspector/SceneInspector.svelte
